Відокремлює функцію createMarkup та деякі правки
