+++ src/js/createMarkup.js
@@ -1,0 +1,70 @@
+import { IMAGE_URL } from "./apiVariables";
+import { getGenreById } from './getGenres';
+
+
+const galleryEl = document.querySelector('.main__movie-card-list');
+
+// Якщо неправильний пошук
+// const headerContainerEl = document.querySelector('.header-container');
+// const failedSearch = document.createElement("p");
+// failedSearch.classList.add("header__failed-search-msg");
+// failedSearch.textContent = "Search result not successful. Enter the correct movie name";
+const failedSearch = document.querySelector('.header__failed-search-msg');
+
+export function createMarkup(moviesArray) {
+    if (moviesArray.length === 0) {
+        failedSearch.style.display = "block";
+    }
+    else {
+        failedSearch.style.display = "none";
+    }
+
+    // failedSearch.innerHTML("");
+
+    const markup = moviesArray.map(movie => {
+        console.log(movie);
+        const posterPath = movie.backdrop_path;
+        let posterPicture = `${IMAGE_URL + posterPath}`;
+        if (posterPath === null) {
+            posterPicture = "https://www.prokerala.com/movies/assets/img/no-poster-available.jpg";
+        } 
+
+        let movieYear = movie.release_date.slice(0, 4);
+        if (movie.release_date === "") {
+            movieYear = "Unknown"
+        }
+
+        const movieTitle = movie.title;
+          
+
+        // const genreIds = movie.genre_ids;
+        // const genresArray = genreIds.map(genreId => {
+        //     return getGenreById(genreId);
+        // });
+
+        // console.log(genresArray);
+        // let genreList = "";
+
+        // if (genresArray.length > 2) {
+        //     const [firstGenre, secondGenre, ...othersGenres] = genresArray;
+
+        //     genreList = `${firstGenre.charAt(0).toUpperCase() + firstGenre.slice(1)}, ${secondGenre.charAt(0).toUpperCase() + secondGenre.slice(1)}, Other`;
+            
+        // }
+
+
+        const movieRaiting = movie.vote_average;
+
+        return `
+                <li class="main__movie-card-item">
+                    <img class="main__movie-img" src="${posterPicture}" alt="Movie poster">
+                    <div class="main__movie-info">
+                        <h2 class="main__movie-title">${movieTitle}</h2>
+                        <p class="main__movie-genre">CAAAAAAT<span class="main_movie-year">${movieYear}</span></p>
+                        <p class="main__movie-raiting">${movieRaiting}</p>
+                    </div>
+                </li>
+        `}).join('');
+        
+        galleryEl.insertAdjacentHTML('beforeend', markup);
+}
