+++ src/partials/header-main.html
@@ -31,5 +31,8 @@
                 </svg>
             </button>
         </form>
+        <p class="header__failed-search-msg">
+            Search result not successful. Enter the correct movie name
+        </p>
     </div>
 </header>
