+++ src/sass/components/_header-main.scss
@@ -136,15 +136,13 @@
 }
 
 .header__failed-search-msg {
+  display: none;
   color: #FF001B;
   font-weight:  400;
   font-size: 14px; 
   line-height: 1.14;
   
-
   @media screen and (min-width: 768px) {
     text-align: center;
   }
-
-  
 }
